feat(ui): add network icon overlay
